requirements.txt
# 知旅(TravelDNA)项目统一依赖文件
# 包含Agent模块和Community模块的所有依赖

# ==================== 核心框架 ====================
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-cors==4.0.0

# ==================== 数据库相关 ====================
pymysql==1.1.2
sqlalchemy==2.0.23

# ==================== 环境变量管理 ====================
python-dotenv==1.0.0

# ==================== HTTP请求 ====================
requests>=2.31.0
aiohttp==3.8.5

# ==================== AI模型API ====================
openai>=1.0.0  # DeepSeek API使用OpenAI兼容格式

# ==================== 数据处理 ====================
numpy==1.24.3
pandas>=2.0.0
faker==19.6.2
orjson==3.9.7

# ==================== 日期时间处理 ====================
python-dateutil==2.8.2

# ==================== 异步支持 ====================
aiofiles==23.2.1

# ==================== 缓存 ====================
cachetools==5.3.1

# ==================== 日志 ====================
loguru==0.7.2

# ==================== 配置验证 ====================
pydantic==2.4.2

# ==================== 文本处理 ====================
jieba==0.42.1

# ==================== 数学计算 ====================
scipy==1.11.3

# ==================== 开发和测试 ====================
pytest==7.4.2
pytest-asyncio==0.21.1

# ==================== Web爬虫相关 ====================
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
fake-useragent>=1.4.0
openpyxl>=3.1.0

# ==================== 说明 ====================
# Agent模块：智能旅游对话系统
# - 需要requests, numpy, pandas等用于API调用和数据处理
# - 需要jieba用于中文文本处理
# - 需要aiohttp, aiofiles用于异步处理
# - 需要beautifulsoup4, selenium等用于网页爬虫
#
# Community模块：社区匹配系统  
# - 需要flask-sqlalchemy, pymysql用于数据库操作
# - 需要faker用于生成测试用户数据
#
# 安装方法：
# pip install -r requirements.txt
